Fixed extra lines data can be used in bar chart
