+++ lib/src/chart/bar_chart/bar_chart_data.dart
@@ -48,6 +48,7 @@ class BarChartData extends AxisChartData with EquatableMixin {
     double? groupsSpace,
     BarChartAlignment? alignment,
     FlTitlesData? titlesData,
+    ExtraLinesData? extraLinesData,
     BarTouchData? barTouchData,
     double? maxY,
     double? minY,
@@ -71,6 +72,7 @@ class BarChartData extends AxisChartData with EquatableMixin {
               ),
           gridData: gridData ?? FlGridData(),
           borderData: borderData,
+          extraLinesData: extraLinesData,
           rangeAnnotations: rangeAnnotations ?? RangeAnnotations(),
           backgroundColor: backgroundColor,
           touchData: barTouchData ?? BarTouchData(),
@@ -92,6 +94,7 @@ class BarChartData extends AxisChartData with EquatableMixin {
     FlTitlesData? titlesData,
     RangeAnnotations? rangeAnnotations,
     BarTouchData? barTouchData,
+    ExtraLinesData? extraLinesData,
     FlGridData? gridData,
     FlBorderData? borderData,
     double? maxY,
@@ -108,6 +111,7 @@ class BarChartData extends AxisChartData with EquatableMixin {
       barTouchData: barTouchData ?? this.barTouchData,
       gridData: gridData ?? this.gridData,
       borderData: borderData ?? this.borderData,
+      extraLinesData: extraLinesData ?? this.extraLinesData,
       maxY: maxY ?? this.maxY,
       minY: minY ?? this.minY,
       baselineY: baselineY ?? this.baselineY,
@@ -127,6 +131,8 @@ class BarChartData extends AxisChartData with EquatableMixin {
         rangeAnnotations:
             RangeAnnotations.lerp(a.rangeAnnotations, b.rangeAnnotations, t),
         barTouchData: b.barTouchData,
+        extraLinesData:
+            ExtraLinesData.lerp(a.extraLinesData, b.extraLinesData, t),
         gridData: FlGridData.lerp(a.gridData, b.gridData, t),
         borderData: FlBorderData.lerp(a.borderData, b.borderData, t),
         maxY: lerpDouble(a.maxY, b.maxY, t),
@@ -146,6 +152,7 @@ class BarChartData extends AxisChartData with EquatableMixin {
         groupsSpace,
         alignment,
         titlesData,
+        extraLinesData,
         barTouchData,
         maxY,
         minY,
