+++ lib/src/chart/line_chart/line_chart_data.dart
@@ -20,9 +20,6 @@ class LineChartData extends AxisChartData with EquatableMixin {
   /// Fills area between two [LineChartBarData] with a color or gradient.
   final List<BetweenBarsData> betweenBarsData;
 
-  /// [LineChart] draws some horizontal or vertical lines on above or below of everything
-  final ExtraLinesData extraLinesData;
-
   /// Handles touch behaviors and responses.
   final LineTouchData lineTouchData;
 
@@ -78,13 +75,13 @@ class LineChartData extends AxisChartData with EquatableMixin {
     Color? backgroundColor,
   })  : lineBarsData = lineBarsData ?? const [],
         betweenBarsData = betweenBarsData ?? const [],
-        extraLinesData = extraLinesData ?? ExtraLinesData(),
         lineTouchData = lineTouchData ?? LineTouchData(),
         showingTooltipIndicators = showingTooltipIndicators ?? const [],
         super(
           gridData: gridData ?? FlGridData(),
           touchData: lineTouchData ?? LineTouchData(),
           borderData: borderData,
+          extraLinesData: extraLinesData,
           titlesData: titlesData ?? FlTitlesData(),
           rangeAnnotations: rangeAnnotations ?? RangeAnnotations(),
           clipData: clipData ?? FlClipData.none(),
@@ -119,8 +116,6 @@ class LineChartData extends AxisChartData with EquatableMixin {
         backgroundColor: Color.lerp(a.backgroundColor, b.backgroundColor, t),
         borderData: FlBorderData.lerp(a.borderData, b.borderData, t),
         clipData: b.clipData,
-        extraLinesData:
-            ExtraLinesData.lerp(a.extraLinesData, b.extraLinesData, t),
         gridData: FlGridData.lerp(a.gridData, b.gridData, t),
         titlesData: FlTitlesData.lerp(a.titlesData, b.titlesData, t),
         rangeAnnotations:
@@ -163,7 +158,6 @@ class LineChartData extends AxisChartData with EquatableMixin {
       betweenBarsData: betweenBarsData ?? this.betweenBarsData,
       titlesData: titlesData ?? this.titlesData,
       rangeAnnotations: rangeAnnotations ?? this.rangeAnnotations,
-      extraLinesData: extraLinesData ?? this.extraLinesData,
       lineTouchData: lineTouchData ?? this.lineTouchData,
       showingTooltipIndicators:
           showingTooltipIndicators ?? this.showingTooltipIndicators,
@@ -1035,287 +1029,6 @@ bool showAllDots(FlSpot spot, LineChartBarData barData) {
   return true;
 }
 
-/// Holds data for drawing extra horizontal lines.
-///
-/// [LineChart] draws some [HorizontalLine] (set by [LineChartData.extraLinesData]),
-/// in below or above of everything, it draws from left to right side of the chart.
-class HorizontalLine extends FlLine with EquatableMixin {
-  /// Draws from left to right of the chart using the [y] value.
-  final double y;
-
-  /// Use it for any kind of image, to draw it in left side of the chart.
-  Image? image;
-
-  /// Use it for vector images, to draw it in left side of the chart.
-  SizedPicture? sizedPicture;
-
-  /// Draws a text label over the line.
-  final HorizontalLineLabel label;
-
-  /// [LineChart] draws horizontal lines from left to right side of the chart
-  /// in the provided [y] value, and color it using [color].
-  /// You can define the thickness using [strokeWidth]
-  ///
-  /// It draws a [label] over it.
-  ///
-  /// You can have a dashed line by filling [dashArray] with dash size and space respectively.
-  ///
-  /// It draws an image in left side of the chart, use [sizedPicture] for vectors,
-  /// or [image] for any kind of image.
-  HorizontalLine({
-    required this.y,
-    HorizontalLineLabel? label,
-    Color? color,
-    double? strokeWidth,
-    List<int>? dashArray,
-    this.image,
-    this.sizedPicture,
-  })  : label = label ?? HorizontalLineLabel(),
-        super(
-            color: color ?? Colors.black,
-            strokeWidth: strokeWidth ?? 2,
-            dashArray: dashArray);
-
-  /// Lerps a [HorizontalLine] based on [t] value, check [Tween.lerp].
-  static HorizontalLine lerp(HorizontalLine a, HorizontalLine b, double t) {
-    return HorizontalLine(
-      y: lerpDouble(a.y, b.y, t)!,
-      label: HorizontalLineLabel.lerp(a.label, b.label, t),
-      color: Color.lerp(a.color, b.color, t),
-      strokeWidth: lerpDouble(a.strokeWidth, b.strokeWidth, t),
-      dashArray: lerpIntList(a.dashArray, b.dashArray, t),
-      image: b.image,
-      sizedPicture: b.sizedPicture,
-    );
-  }
-
-  /// Used for equality check, see [EquatableMixin].
-  @override
-  List<Object?> get props => [
-        y,
-        label,
-        color,
-        strokeWidth,
-        dashArray,
-        image,
-        sizedPicture,
-      ];
-}
-
-/// Holds data for drawing extra vertical lines.
-///
-/// [LineChart] draws some [VerticalLine] (set by [LineChartData.extraLinesData]),
-/// in below or above of everything, it draws from bottom to top side of the chart.
-class VerticalLine extends FlLine with EquatableMixin {
-  /// Draws from bottom to top of the chart using the [x] value.
-  final double x;
-
-  /// Use it for any kind of image, to draw it in bottom side of the chart.
-  Image? image;
-
-  /// Use it for vector images, to draw it in bottom side of the chart.
-  SizedPicture? sizedPicture;
-
-  /// Draws a text label over the line.
-  final VerticalLineLabel label;
-
-  /// [LineChart] draws vertical lines from bottom to top side of the chart
-  /// in the provided [x] value, and color it using [color].
-  /// You can define the thickness using [strokeWidth]
-  ///
-  /// It draws a [label] over it.
-  ///
-  /// You can have a dashed line by filling [dashArray] with dash size and space respectively.
-  ///
-  /// It draws an image in bottom side of the chart, use [sizedPicture] for vectors,
-  /// or [image] for any kind of image.
-  VerticalLine({
-    required this.x,
-    VerticalLineLabel? label,
-    Color? color,
-    double? strokeWidth,
-    List<int>? dashArray,
-    this.image,
-    this.sizedPicture,
-  })  : label = label ?? VerticalLineLabel(),
-        super(
-            color: color ?? Colors.black,
-            strokeWidth: strokeWidth ?? 2,
-            dashArray: dashArray);
-
-  /// Lerps a [VerticalLine] based on [t] value, check [Tween.lerp].
-  static VerticalLine lerp(VerticalLine a, VerticalLine b, double t) {
-    return VerticalLine(
-      x: lerpDouble(a.x, b.x, t)!,
-      label: VerticalLineLabel.lerp(a.label, b.label, t),
-      color: Color.lerp(a.color, b.color, t),
-      strokeWidth: lerpDouble(a.strokeWidth, b.strokeWidth, t),
-      dashArray: lerpIntList(a.dashArray, b.dashArray, t),
-      image: b.image,
-      sizedPicture: b.sizedPicture,
-    );
-  }
-
-  /// Used for equality check, see [EquatableMixin].
-  @override
-  List<Object?> get props => [
-        x,
-        label,
-        color,
-        strokeWidth,
-        dashArray,
-        image,
-        sizedPicture,
-      ];
-}
-
-/// Shows a text label
-abstract class FlLineLabel with EquatableMixin {
-  /// Determines showing label or not.
-  final bool show;
-
-  /// Inner spaces around the drawing text.
-  final EdgeInsetsGeometry padding;
-
-  /// Sets style of the drawing text.
-  final TextStyle? style;
-
-  /// Aligns the text on the line.
-  final Alignment alignment;
-
-  /// Draws a title on the line, align it with [alignment] over the line,
-  /// applies [padding] for spaces, and applies [style] for changing color,
-  /// size, ... of the text.
-  /// [show] determines showing label or not.
-  FlLineLabel(
-      {required this.show,
-      required this.padding,
-      required this.style,
-      required this.alignment});
-
-  /// Used for equality check, see [EquatableMixin].
-  @override
-  List<Object?> get props => [
-        show,
-        padding,
-        style,
-        alignment,
-      ];
-}
-
-/// Draws a title on the [HorizontalLine]
-class HorizontalLineLabel extends FlLineLabel with EquatableMixin {
-  /// Resolves a label for showing.
-  final String Function(HorizontalLine) labelResolver;
-
-  /// Returns the [HorizontalLine.y] as the drawing label.
-  static String defaultLineLabelResolver(HorizontalLine line) =>
-      line.y.toStringAsFixed(1);
-
-  /// Draws a title on the [HorizontalLine], align it with [alignment] over the line,
-  /// applies [padding] for spaces, and applies [style for changing color,
-  /// size, ... of the text.
-  /// Drawing text will retrieve through [labelResolver],
-  /// you can override it with your custom data.
-  /// /// [show] determines showing label or not.
-  HorizontalLineLabel({
-    EdgeInsets? padding,
-    TextStyle? style,
-    Alignment? alignment,
-    bool show = false,
-    String Function(HorizontalLine)? labelResolver,
-  })  : labelResolver =
-            labelResolver ?? HorizontalLineLabel.defaultLineLabelResolver,
-        super(
-          show: show,
-          padding: padding ?? const EdgeInsets.all(6),
-          style: style,
-          alignment: alignment ?? Alignment.topLeft,
-        );
-
-  /// Lerps a [HorizontalLineLabel] based on [t] value, check [Tween.lerp].
-  static HorizontalLineLabel lerp(
-      HorizontalLineLabel a, HorizontalLineLabel b, double t) {
-    return HorizontalLineLabel(
-      padding:
-          EdgeInsets.lerp(a.padding as EdgeInsets, b.padding as EdgeInsets, t),
-      style: TextStyle.lerp(a.style, b.style, t),
-      alignment: Alignment.lerp(a.alignment, b.alignment, t),
-      labelResolver: b.labelResolver,
-      show: b.show,
-    );
-  }
-
-  /// Used for equality check, see [EquatableMixin].
-  @override
-  List<Object?> get props => [
-        labelResolver,
-        show,
-        padding,
-        style,
-        alignment,
-      ];
-}
-
-/// Draws a title on the [VerticalLine]
-class VerticalLineLabel extends FlLineLabel with EquatableMixin {
-  /// Resolves a label for showing.
-  final String Function(VerticalLine) labelResolver;
-
-  /// Returns the [VerticalLine.x] as the drawing label.
-  static String defaultLineLabelResolver(VerticalLine line) =>
-      line.x.toStringAsFixed(1);
-
-  /// Draws a title on the [VerticalLine], align it with [alignment] over the line,
-  /// applies [padding] for spaces, and applies [style for changing color,
-  /// size, ... of the text.
-  /// Drawing text will retrieve through [labelResolver],
-  /// you can override it with your custom data.
-  /// [show] determines showing label or not.
-  VerticalLineLabel({
-    EdgeInsets? padding,
-    TextStyle? style,
-    Alignment? alignment,
-    bool? show,
-    String Function(VerticalLine)? labelResolver,
-  })  : labelResolver =
-            labelResolver ?? VerticalLineLabel.defaultLineLabelResolver,
-        super(
-          show: show ?? false,
-          padding: padding ?? const EdgeInsets.all(6),
-          style: style ??
-              const TextStyle(
-                color: Colors.black,
-                fontWeight: FontWeight.bold,
-                fontSize: 14,
-              ),
-          alignment: alignment ?? Alignment.bottomRight,
-        );
-
-  /// Lerps a [VerticalLineLabel] based on [t] value, check [Tween.lerp].
-  static VerticalLineLabel lerp(
-      VerticalLineLabel a, VerticalLineLabel b, double t) {
-    return VerticalLineLabel(
-      padding:
-          EdgeInsets.lerp(a.padding as EdgeInsets, b.padding as EdgeInsets, t),
-      style: TextStyle.lerp(a.style, b.style, t),
-      alignment: Alignment.lerp(a.alignment, b.alignment, t),
-      labelResolver: b.labelResolver,
-      show: b.show,
-    );
-  }
-
-  /// Used for equality check, see [EquatableMixin].
-  @override
-  List<Object?> get props => [
-        labelResolver,
-        show,
-        padding,
-        style,
-        alignment,
-      ];
-}
-
 /// Holds data for showing a vector image inside the chart.
 ///
 /// for example:
@@ -1354,47 +1067,6 @@ class SizedPicture with EquatableMixin {
         picture,
         width,
         height,
-      ];
-}
-
-/// Draws some straight horizontal or vertical lines in the [LineChart]
-class ExtraLinesData with EquatableMixin {
-  final List<HorizontalLine> horizontalLines;
-  final List<VerticalLine> verticalLines;
-
-  final bool extraLinesOnTop;
-
-  /// [LineChart] draws some straight horizontal or vertical lines,
-  /// you should set [LineChartData.extraLinesData].
-  /// Draws horizontal lines using [horizontalLines],
-  /// and vertical lines using [verticalLines].
-  ///
-  /// If [extraLinesOnTop] sets true, it draws the line above the main bar lines, otherwise
-  /// it draws them below the main bar lines.
-  ExtraLinesData({
-    List<HorizontalLine>? horizontalLines,
-    List<VerticalLine>? verticalLines,
-    bool? extraLinesOnTop,
-  })  : horizontalLines = horizontalLines ?? const [],
-        verticalLines = verticalLines ?? const [],
-        extraLinesOnTop = extraLinesOnTop ?? true;
-
-  /// Lerps a [ExtraLinesData] based on [t] value, check [Tween.lerp].
-  static ExtraLinesData lerp(ExtraLinesData a, ExtraLinesData b, double t) {
-    return ExtraLinesData(
-      extraLinesOnTop: b.extraLinesOnTop,
-      horizontalLines:
-          lerpHorizontalLineList(a.horizontalLines, b.horizontalLines, t),
-      verticalLines: lerpVerticalLineList(a.verticalLines, b.verticalLines, t),
-    );
-  }
-
-  /// Used for equality check, see [EquatableMixin].
-  @override
-  List<Object?> get props => [
-        horizontalLines,
-        verticalLines,
-        extraLinesOnTop,
       ];
 }
